feat: Add isDarkImage property to Technology type and update Technology component
The code changes include adding a new property `isDarkImage` to the `Technology` type in the `Technology.ts` file. This property will be used to determine if the image for a technology should be displayed in a dark mode. Additionally, the `Technology` component's styles in the `technology.component.scss` file have been updated to adjust the contrast of the background image.

Recent user commits:
- Update Go backend technology image and styles
- Update Go backend technology name to "Go" in projects.json
- Update Go backend technology name and image
- Update project dependencies and styles
- Add portfolio projects and technologies data
- Update project dependencies and styles